% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler}
\alias{sampler}
\title{Generate Sample Lists from Excel or CSV Files}
\arguments{
\item{ci}{the required confidence level}

\item{me}{the margin of error}

\item{p}{the expected probability of occurrence}

\item{backups}{the number of available replacements}

\item{seed}{the random number seed}
}
\value{
Writes samples to an Excel workbook and generates a report summary.
}
\description{
\code{sampler} generates Simple Random or Stratified samples
}
\section{Details}{

\code{sampler} lets users select an Excel or delimited text (.csv or .txt) data file and the type of sample they prefer (Simple Random Sample, Stratified Random Sample, or Tabbed Stratified Sample with each stratum in a different Excel worksheet).
}

\examples{
if(interactive()){
sampler(backups=3, p=0.6)
}
}
