% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler}
\alias{sampler}
\title{Generate Sample Lists from Excel or CSV Files}
\usage{
sampler()
}
\value{
Writes samples to an Excel workbook
}
\description{
\code{sampler} generates Simple Random or Stratified samples
}
